Avec la distance euclidienne ça marche mieux con que je suis

+++ Core/AbstractObjects/ISolution.cs
@@ -3,7 +3,7 @@
 public interface ISolution
 {
     
-    float GetFitness();
+    double GetFitness();
     List<ISolution> GetNeighbours();
     
     void GenerateRandomSolution();
